feat: add institution detail page

+++ backend/app/Http/Controllers/DonationController.php
@@ -5,34 +5,44 @@ namespace App\Http\Controllers;
 use App\Http\Requests\Donation\DonationStoreRequest;
 use App\Http\Requests\Donation\DonationUpdateRequest;
 use App\Models\Donation;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class DonationController extends Controller
 {
-    public function index()
+    public function index(Request $request)
     {
-        $donations = Donation::all();
+        /** @var \App\Models\User $user **/
+        $user = $request->user();
+
+        $donations = $user->donations()->get();
         return response()->json($donations);
     }
 
     public function store(DonationStoreRequest $request)
     {
-        $donation = Donation::create($request->validated());
+        /** @var \App\Models\User $user **/
+        $user = $request->user();
+        $donation = $user->donations()->create($request->validated());
         return response()->json($donation);
     }
 
     public function show(Donation $donation)
     {
+        Gate::authorize('view', $donation);
         return response()->json($donation);
     }
 
     public function update(DonationUpdateRequest $request, Donation $donation)
     {
+        Gate::authorize('update', $donation);
         $donation->update($request->validated());
         return response()->json($donation);
     }
 
     public function destroy(Donation $donation)
     {
+        Gate::authorize('delete', $donation);
         $donation->delete();
         return response(null, 204);
     }
